+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.18.0] - 2019-02-21
 ### Added
 - VIMEET-2011 - Prioritize meeting request before planning
 - VIMEET-2047 - Add number max of meetings per sheet in participation type parameter